+++ cmd/constants/constants.go
@@ -1,0 +1,3 @@
+package constants
+
+var AvailableGroups = []string{"A", "B", "C", "D", "E", "F", "G", "H", "a", "b", "c", "d", "e","f","g","h"}
